% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolphins.R
\docType{data}
\name{dolphins}
\alias{dolphins}
\title{Social network of bottlenose dolphins}
\format{
An object of class \code{dgCMatrix} with 62 rows and 62 columns.
}
\source{
\href{http://www-personal.umich.edu/~mejn/netdata/}{http://www-personal.umich.edu/~mejn/netdata/}
}
\usage{
data(dolphins)
}
\description{
Social network of bottlenose dolphins
}
\references{
Jerome Kunegis. KONECT - The Koblenz Network Collection. In Proc. Int. Conf. on World Wide Web Companion, pages 1343-1350, 2013
(\href{https://dl.acm.org/doi/10.1145/2487788.2488173}{http})

D. Lusseau, K. Schneider, O. J. Boisseau, P. Haase, E. Slooten, and S. M. Dawson. The bottlenose dolphin community of Doubtful Sound features a large proportion of long-lasting associations. Behav. Ecol. and Sociobiol., 54:396-405, 2003.
}
\keyword{datasets}
