% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWTest.R
\name{sw_test}
\alias{sw_test}
\title{sw_test}
\usage{
sw_test(graph, baseline.model, num.iter)
}
\arguments{
\item{graph}{An input network}

\item{baseline.model}{Model to test against}

\item{num.iter}{Number of bootstrap iterations}
}
\value{
Prints result of the test
}
\description{
Performs a hypothesis test for the network small-world property
}
\examples{
data(dolphins)
sw_test(dolphins, "ER", 100)
}
