% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWTest.R
\name{sw_test}
\alias{sw_test}
\title{sw_test}
\usage{
sw_test(x, model.list, num.iter)
}
\arguments{
\item{x}{A square adjacency matrix.}

\item{model.list}{Vector of baseline models to test against. Possible values are ER, SBM, DCSBM and CL.}

\item{num.iter}{Vector of number of bootstrap iterations for each baseline model.}
}
\value{
Prints result of the test
}
\description{
Performs a hypothesis test for the network small-world property
}
\examples{
data(dolphins)
sw_test(dolphins, c("ER", "SBM"), c(100, 50))
}
